ME QUIERO SUICIDAR

+++ stats/src/main/java/org/proyecto/DatosSalidaJSON.java
@@ -1,6 +1,8 @@
 package org.proyecto;
 
+import java.math.BigInteger;
+
 public record DatosSalidaJSON(double[] dunnIndex, double[] averageIndex, double averageTotalIndex, double randIndex,
                               double wcss, int clusters, double bcss, double CHIndex, double DBIndex,
-                              double averageSilhouette, long elapsedTime) {
+                              double averageSilhouette, BigInteger elapsedTime) {
 }
